The QueryController is no longer coupled with the QueryControllerTreeModel. Adding listeners should be occurred in the client-side code (i.e., SavedQueriesPanel).
There have been many changes for the decoupling process:
- The QueryControllerTreeModel object was deleted from the QueryController class to
remove the dependency, including the getTreeModel() method as well.
- The QueryController object was deleted from the QueryControllerTreeModel constructor
to remove the dependency. The old init() method was renamed into synchronize(Vector)
with some slight changes in the method body.
- The SavedQueriesPanel is not necessary to implement the query listener. The listener
is already implemented by QueryControllerTreeModel and, thus, we just need to register
it to the QueryController.
- The synchronize(Vector) method was used to implement the refresh tree model function.