- The OBDAPreference object is no longer using singleton pattern. The only thing that remains single is the "defaultValue" HashMap that holds (well) the default values. These values are loaded first (hard-coded) and we store them in a static object.
- As the consequence of eliminating the singleton pattern, the OBDAPreference
object has to be passed to all other objects that need to read the preference
strings, e.g., the Preference panel or the Mapping panel.

- The initialization of the OBDAPreference object happens when Protege is
started and we synchronize it to the EditorKit. Note: IMO, it is not
appropriate to set it there due to the fact EditorKit is not intended to 
hold preferences. Finding a better solution is strongly suggested and can be 
helpful to increase the quality of the API.

- We also removed several unnecessary static methods in the OBDAPreference
class.

- We fix the bug in the Preference panel for the highlight color preview. It
seems in the MacOS environment, we have to call setOpaque(true) after setting
the JButton background.

- We add meaningful variable names in the Preference panel GUI to increase
its readability. We fix also the layout manager to be more precise on each
JComponent measure.