Improving the mapping filter feature.
There are three major changes in the code. The first is the introduction of a MappingFilter grammar. The grammar will help us to disseminate the filter string and to construct the filter object accordingly. Second, we add the "NOT" keyword in the grammar so that a user can specify a negation in his search. And lastly, we did also some code clean up in the old filter classes to integrate with the grammar module.

BUG=37

Ticket URL=https://babbage.inf.unibz.it/trac/obdapublic/ticket/37